LayoutTests/http/tests/referrer-policy-img/unsafe-url/cross-origin-http-UpgradeMixedContent.https.html

<!DOCTYPE html><!-- webkit-test-runner [ UpgradeMixedContentEnabled=true ] -->
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests the behavior of unsafe-url referrer policy when cross origin from HTTPS to HTTP.");
jsTestIsAsync = true;

if (window.internals) {
    internals.settings.setAllowDisplayOfInsecureContent(true);
    internals.setTrackingPreventionEnabled(false);
}

if (window.testRunner) {
    testRunner.setStatisticsShouldDowngradeReferrer(false, function () {
        // Should be the full URL
        let expected = document.location.href
        let img = new Image();
        img.referrerPolicy = "unsafe-url";
        img.onload = function() { testPassed('referrer is "' + expected + '"'); finishJSTest(); }
        img.onerror = function() { testFailed("referrer is not unsafe-url"); finishJSTest(); }
        img.src = "http://localhost:8000/referrer-policy/resources/image.py?expected=" + expected;
        if (window.internals)
            internals.setTrackingPreventionEnabled(true);
    });
}

</script>
</body>
</html>
